# Faster itertoolz.update_in