@groutr 2018-12-18
@eriknw do you have any feedback on this change?