@eriknw 2019-06-22
LGTM, thanks @groutr!  Are there any other speed improvements to steal from `cytoolz`?

Also, aesthetically, this uses more blank lines then I generally prefer, but I'd like to get this in, so merging!