@groutr 2019-06-24
@eriknw I'm sure there are probably other speed improvements that could be carried over from cytoolz, but I haven't looked in awhile.  Thanks for merging!